OrIConversion Trait, View-ability of Ori field, and Rotation traits update
Made all of the orientation conversions into a set of traits that now
all of the orientations implement.

The ori field within each orientation struct is no longer public. It is
now preferable for new orientation structs to be created using either
the new or new_init function. In order to view or obtain a mutable view
of the ori field one must call the ori_view or ori_view_mut functions.

Finally, I’ve added a set of traits for rotating either vectors or
tensor fields. Currently, the only orientation field that supports
these new traits is the rotation matrix.

The library can now do some useful things with the different
orientations besides just being able to convert between one or the
other. It is slowly inching closer towards a public 0.1.0 release on
Crates.io. Documentations, test examples, and a few unimplemented trait
features still stand in the way.