Refactor for more direct orientation conversions
Refactored the code such that more direct conversions exist. It's much less likely now for the internal structure to have to convert between to one structure to another before it is able to reach the desired orientation structure.

Most structures had large changes made to them. I however did not do this with the compact Rodrigues vector representation. It has the possibility of being badly behaved, so I decided to limit the number of direct conversions it had available to it.