Transpose and conjugate operations
For Rotation matrices, axis-angle variants, and Rodrigues vector variants the transpose operation was added which operates on each element.

Quaternions had the conjugate operation added as well which operates on each element.

The equivalent inplace operations were also added that perform each of the above operations on the internal ori field of the structure for each orientation operation.

Other orientations did not have the above added because there is no easy equivalent operation for any of them.

Finally, a few unit tests should eventually be tested to make sure that they work.