Quaternion product
Added the quaternion product to the methods of Quat. It is with commit that I would say that all of the main quaternion functions are represented. I would say at this point we can now add parallel capabilities to the library. Since, we now have all of the main functionalities that I set out for the intial cargo release.