Initial parallelization trials
I've got an initial version of the parallel code in here.
It unfortunately seems to have led to a huge increase in compilation time. I think I'm going to have to split the library into essentially two crates: one that's serial and one that's parallel. It will lead to code duplication unfortunately.